fix metadata versions being out-of-sync
