+++ secureli/services/updater.py
@@ -4,6 +4,7 @@ import pydantic
 
 from secureli.abstractions.pre_commit import PreCommitAbstraction
 from secureli.repositories.secureli_config import SecureliConfigRepository
+from secureli.services.language_support import LanguageSupportService
 
 
 class UpdateResult(pydantic.BaseModel):
@@ -22,10 +23,10 @@ class UpdaterService:
 
     def __init__(
         self,
-        pre_commit: PreCommitAbstraction,
+        language_support: LanguageSupportService,
         config: SecureliConfigRepository,
     ):
-        self.pre_commit = pre_commit
+        self.language_support = language_support
         self.config = config
 
     def update_hooks(
@@ -43,14 +44,16 @@ class UpdaterService:
         :param repos: Dectionary of repos to update. This is used to target specific repos instead of all repos.
         :return: ExecuteResult, indicating success or failure.
         """
-        update_result = self.pre_commit.autoupdate_hooks(bleeding_edge, freeze, repos)
+        update_result = self.language_support.autoupdate_hooks(
+            bleeding_edge, freeze, repos
+        )
         output = update_result.output
 
         if update_result.successful and not output:
             output = "No changes necessary.\n"
 
         if update_result.successful and update_result.output:
-            prune_result = self.pre_commit.remove_unused_hooks()
+            prune_result = self.language_support.remove_unused_hooks()
             output = output + "\nRemoving unused environments:\n" + prune_result.output
 
         return UpdateResult(successful=update_result.successful, output=output)
@@ -62,12 +65,14 @@ class UpdaterService:
         """
         secureli_config = self.config.load()
         output = "Updating .pre-commit-config.yaml...\n"
-        install_result = self.pre_commit.install(language=secureli_config.languages)
-        if not install_result.successful:
+        install_result = self.language_support.apply_support(
+            languages=secureli_config.languages
+        )
+        if not install_result.version:
             output += "Failed to update .pre-commit-config.yaml prior to hook install\n"
-            return UpdateResult(successful=install_result.successful, output=output)
+            return UpdateResult(successful=False, output=output)
 
-        hook_install_result = self.pre_commit.update()
+        hook_install_result = self.language_support.update()
         output += hook_install_result.output
 
         if (
@@ -77,7 +82,7 @@ class UpdaterService:
             output += "No changes necessary.\n"
 
         if hook_install_result.successful and hook_install_result.output:
-            prune_result = self.pre_commit.remove_unused_hooks()
+            prune_result = self.language_support.remove_unused_hooks()
             output += "\nRemoving unused environments:\n" + prune_result.output
 
         return UpdateResult(successful=hook_install_result.successful, output=output)
